% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_gaussian_to_sf}
\alias{add_gaussian_to_sf}
\title{add_guassian_to_sf}
\usage{
add_gaussian_to_sf(sf_obj, lat_0, lon_0, ws_mps, wd_deg, x_0_m = 1000)
}
\arguments{
\item{sf_obj}{An sf object with at least one numeric column.}

\item{lat_0}{gaussian plume source latitude in degrees}

\item{lon_0}{guassian plume source longitude in degrees}

\item{ws_mps}{local surface windspeed in meters per second}

\item{wd_deg}{local surface wind direction in degrees}

\item{x_0_m}{guassian plume length scale in meters (default 1 km)}
}
\value{
original sf_obj with guassian plume model eastimate at each geometry
}
\description{
add_guassian_to_sf
}
\examples{
grid <- add_gaussian_to_sf(target_grid, lat_0=19.5, lon_0=53.3, ws_mps=2.5, wd_deg =270, x_0_m=1000)
}
