% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{buffered_convex_hull}
\alias{buffered_convex_hull}
\title{Title}
\usage{
buffered_convex_hull(soundings, grid, buffer)
}
\arguments{
\item{soundings}{An sf object (thought to be the original soundings)}

\item{grid}{An sf object containing the underlying fine grid.}

\item{buffer}{A numeric value specifying a buffer to put around the soundings (meters)}
}
\value{
original grid with added cells to increase bounding box beyond soundings bounding box
}
\description{
Title
}
\examples{
buffered_grid <- buffered_convex_hull(soundings, target_grid, buffer=500)
}
