% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{build_region_list}
\alias{build_region_list}
\title{Build region list from observation and (optional) stratum polygons}
\usage{
build_region_list(
  obs_polygons,
  stratum_polygons = NULL,
  method = c("triangulate", "intersect-then-triangulate", "raw")
)
}
\arguments{
\item{obs_polygons}{An sf object of observation polygons}

\item{stratum_polygons}{Optional sf object of stratifying polygons}

\item{method}{One of "raw", "triangulate", or "intersect-then-triangulate"}
}
\value{
A list of region_composite or region_atomic objects (some may be NULL)
}
\description{
Build region list from observation and (optional) stratum polygons
}
