% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{composite_region_to_sf}
\alias{composite_region_to_sf}
\title{Convert a composite region into an sf object for visualization}
\usage{
composite_region_to_sf(composite)
}
\arguments{
\item{composite}{A region_composite object}
}
\value{
An sf object containing all atomic parts of the region
}
\description{
Convert a composite region into an sf object for visualization
}
