% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_A_matrix.R
\name{compute_A_matrix_2d_gauss}
\alias{compute_A_matrix_2d_gauss}
\title{compute_A_matrix_2d_gauss}
\usage{
compute_A_matrix_2d_gauss(
  fine_grid,
  soundings,
  major_axis_sigma_scaler = 1,
  minor_axis_sigma_scaler = 1
)
}
\arguments{
\item{fine_grid}{An sf object, a fine grid on which we want to estimate SIF or XCO2}

\item{soundings}{An sf object, The observed soundings}

\item{major_axis_sigma_scaler}{a factor to scale the guassian blur in the major axis direction}

\item{minor_axis_sigma_scaler}{a factor to scale the gaussian blue in the minor axis direction}
}
\value{
Matrix of weights corresponding to 2d guassian centered on each sounding
}
\description{
compute_A_matrix_2d_gauss
}
\examples{
# Note: if(interactive()){} does not run in checks and can temporarily wrap long-running examples.
if(interactive()){A2D <- compute_A_matrix_2d_gauss(target_grid, soundings)}
}
