% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_A_matrix.R
\name{compute_A_matrix_future}
\alias{compute_A_matrix_future}
\title{compute_A_matrix}
\usage{
compute_A_matrix_future(
  fine_grid,
  soundings,
  intersection = NULL,
  n_workers = future::availableCores(),
  use_progress = TRUE
)
}
\arguments{
\item{fine_grid}{An sf object, a fine grid on which we want to estimate SIF}

\item{soundings}{An sf object, The observed soundings}
}
\value{
Matrix of wieghted area overlap between grid and soundings
}
\description{
Computes the matrix of weights, denoted $A$, determining overlapping areas.
}
\examples{
if(interactive()){APG <- compute_A_matrix(target_grid, soundings, mc.cores = 2)}
}
