% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_W.R
\name{compute_dominant_class_W_matrix}
\alias{compute_dominant_class_W_matrix}
\title{compute_dominant_class_W_matrix}
\usage{
compute_dominant_class_W_matrix(target_grid, class = 6, epsilon = 0.01)
}
\arguments{
\item{target_grid}{An sf object}

\item{class}{An integer from 1-8. 6 represents water.}

\item{epsilon}{A small nonzero value representing the weight of correlation between pixels with different dominant classes.}
}
\description{
compute_dominant_class_W_matrix
}
