% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{compute_gradient.matern_fourier}
\alias{compute_gradient.matern_fourier}
\title{Compute Gradient of Matérn-Fourier Prior Precision Matrix}
\usage{
\method{compute_gradient}{matern_fourier}(prior)
}
\arguments{
\item{prior}{An object of class "matern_fourier".}
}
\value{
A named list with components:
\describe{
\item{tau}{Derivative w.r.t. \code{tau}.}
\item{kappa}{Derivative w.r.t. \code{kappa}.}
\item{alpha}{Derivative w.r.t. \code{alpha}.}
}
}
\description{
Computes the derivative of the precision matrix with respect to each hyperparameter:
\itemize{
\item \code{tau}: scaling precision.
\item \code{kappa}: range parameter.
\item \code{alpha}: smoothness parameter.
}
All derivatives are returned as sparse diagonal matrices.
}
\details{
The gradients are:
\deqn{
  \frac{\partial Q}{\partial \tau} = \mathrm{diag}\Bigl[\bigl(\kappa^2 + \lambda_j^{Laplacian}\bigr)^{\alpha}\Bigr]
}
\deqn{
  \frac{\partial Q}{\partial \kappa} = \tau \, \mathrm{diag}\Bigl[ \alpha \cdot 2\kappa \cdot \bigl(\kappa^2 + \lambda_j^{Laplacian}\bigr)^{\alpha - 1}\Bigr]
}
\deqn{
  \frac{\partial Q}{\partial \alpha} = \tau \, \mathrm{diag}\Bigl[\bigl(\kappa^2 + \lambda_j^{Laplacian}\bigr)^{\alpha} \cdot \log\bigl(\kappa^2 + \lambda_j^{Laplacian}\bigr)\Bigr]
}
}
