% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{compute_gradient}
\alias{compute_gradient}
\title{Compute Gradient of the Prior Precision Matrix}
\usage{
compute_gradient(prior)
}
\arguments{
\item{prior}{A prior object.}
}
\value{
A named list of sparse matrices: the derivative w.r.t. each hyperparameter.
}
\description{
Compute Gradient of the Prior Precision Matrix
}
