% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{compute_gradient.sar}
\alias{compute_gradient.sar}
\title{Compute Gradient of SAR Prior Precision Matrix}
\usage{
\method{compute_gradient}{sar}(prior)
}
\arguments{
\item{prior}{An object of class "sar".}
}
\value{
A named list of sparse matrices.
}
\description{
Computes the derivative of the precision matrix w.r.t. each hyperparameter.
}
