% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_variance.R
\name{compute_posterior_measurement_error}
\alias{compute_posterior_measurement_error}
\title{compute_posterior_measurement_error}
\usage{
compute_posterior_measurement_error(posterior_mean, X, Lambda, y)
}
\arguments{
\item{posterior_mean}{The posterior mean calculated by fit_spatial_field() or fit_stacked_spatial_field()}

\item{X}{The design matrix computed by a prepare_ type function}

\item{Lambda}{The prior precision matrix before scaling}

\item{y}{A vector of response values}

\item{tau}{The scale factor on the prior precision matrix}
}
\description{
compute_posterior_measurement_error
}
