% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{compute_precision}
\alias{compute_precision}
\title{Compute Precision Matrix for a Prior Object}
\usage{
compute_precision(prior)
}
\arguments{
\item{prior}{A prior object.}
}
\description{
Compute Precision Matrix for a Prior Object
}
