% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{compute_precision.sar}
\alias{compute_precision.sar}
\title{Compute Precision Matrix for SAR Prior}
\usage{
\method{compute_precision}{sar}(prior)
}
\arguments{
\item{prior}{An object of class "sar".}
}
\value{
A sparse symmetric precision matrix (Matrix).
}
\description{
Given a SAR prior object, returns the precision matrix Q.
Reuses cached components when possible.
}
