% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{convert_lat_and_lon_to_aea_crs_string}
\alias{convert_lat_and_lon_to_aea_crs_string}
\title{convert lat and lon to an Alber's Equal Area crs string}
\usage{
convert_lat_and_lon_to_aea_crs_string(lon_0, lat_0, lat_1, lat_2)
}
\arguments{
\item{lon_0}{Longitude of projection center in degrees.}

\item{lat_0}{Latitude of projection center in degrees}

\item{lat_1}{First standard parallel.}

\item{lat_2}{Second standard parallel.}
}
\value{
String appropriate for crs.
}
\description{
convert lat and lon to an Alber's Equal Area crs string
}
\examples{
crs_string <- convert_lat_and_lon_to_aea_crs_string(53.2, 19.5, 19.7, 53.4)
}
