% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_aea_projection_crs_string_at_centroid}
\alias{create_aea_projection_crs_string_at_centroid}
\title{create a crs string corresponding to Alber's Equal Area projection centered
at the centroid of soundings}
\usage{
create_aea_projection_crs_string_at_centroid(
  soundings,
  lat_1 = 29.5,
  lat_2 = 45.5
)
}
\arguments{
\item{soundings}{An sf object (thought to be the original soundings)}
}
\value{
String appropriate for crs caluclated from soundings centroid
}
\description{
create a crs string corresponding to Alber's Equal Area projection centered
at the centroid of soundings
}
\examples{
crs_string <- create_aea_projection_crs_string_at_centroid(soundings, lat_1 = 29.5,lat_2 = 45.5)
}
