% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_timers}
\alias{create_timers}
\title{Title}
\usage{
create_timers(timer_str = "timer0")
}
\arguments{
\item{timer_str}{string value for timeR assignment in global environment}
}
\value{
NULL. timer_str automatically assigned to timeR structure in global environment
}
\description{
Title
}
\examples{
#create_timers(timer_str = "my_timer")
}
