% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{custom_gaussian_mask}
\alias{custom_gaussian_mask}
\title{custom_gaussian_mask}
\usage{
custom_gaussian_mask(
  sf_obj,
  inner_radius_m = 3000,
  outer_radius_m = 30000,
  source_radius_m = 1000,
  gaussian_cutoff_percent = 1,
  gaussian_buffer_percent = 0.1,
  nucleous_cutoff_percent = NULL
)
}
\arguments{
\item{sf_obj}{An sf object containing guassian_gpm2 estimate from add_gaussain_to_sf().}

\item{inner_radius_m}{radius around source to remove from background mask}

\item{outer_radius_m}{radius around source containing valid samples}

\item{source_radius_m}{radius around source to remove from in-plume mask}

\item{gaussian_cutoff_percent}{percent of plume maximum value used as lower limit boundary for in-plume mask}

\item{gaussian_buffer_percent}{percent of plume maximum value used for buffer mask. Must be lower than gaussian_cutoff_percent.}

\item{nucleous_cutoff_percent}{percent of plume maximum value used as upper limit bouardary for in-plume mask (Not Recommended)}
}
\value{
input sf_obj with boolean columns for in-plume, background, and buffer regions of guassian model.
}
\description{
custom_gaussian_mask
}
\examples{
# grid <- custom_gaussian_mask(grid)
}
