% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_integration_rules.R
\name{default_integration_rule}
\alias{default_integration_rule}
\title{Determine Default Integration Rule for a Basis}
\usage{
default_integration_rule(basis)
}
\arguments{
\item{basis}{A basis object (e.g., from make_geometry_basis(), make_matern_fourier_basis(), etc.)}
}
\value{
An integration_rule object
}
\description{
Returns the appropriate integration rule for a given basis type.
}
