% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_rules.R
\name{evaluate_integral.centroid_gaussian}
\alias{evaluate_integral.centroid_gaussian}
\title{Centroid Gaussian integration for a region (ignored)}
\usage{
\method{evaluate_integral}{centroid_gaussian}(rule, region, basis, index)
}
\description{
Centroid Gaussian integration for a region (ignored)
}
