% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_rules.R
\name{evaluate_integral.qmc_average}
\alias{evaluate_integral.qmc_average}
\title{QMC Integration for an Atomic Region (with Optional Sensitivity Weights)}
\usage{
\method{evaluate_integral}{qmc_average}(rule, region, basis, index = NULL, tolerance = 1e-06)
}
\arguments{
\item{rule}{An integration rule object of class "qmc_average".}

\item{region}{A region_atomic object (must have $shape and $coords).}

\item{basis}{A basis object (with $evaluate_fn).}

\item{index}{An integer index of the observation (used to pick sensitivity function).}

\item{tolerance}{A tolerance for sensitivity to set overall weight to 0.}
}
\value{
A numeric vector of length equal to basis$k.
}
\description{
Evaluates the (weighted) average value of basis functions over an atomic region
using quasi-Monte Carlo (QMC) integration.
}
