% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{generate_grid}
\alias{generate_grid}
\title{generate grid}
\usage{
generate_grid(
  soundings,
  res,
  buffer = 2 * res,
  grid_shape = c("buffered_convex_hull", "rectangle"),
  lon_0 = NULL,
  lat_0 = NULL,
  lat_1 = 29.5,
  lat_2 = 45.5
)
}
\arguments{
\item{soundings}{An sf object (thought to be the original soundings)}

\item{res}{A numeric value specifying the resolution (in meters)}

\item{buffer}{A numeric value specifying a buffer to put around the soundings}

\item{grid_shape}{One of "buffered_convex_hull" or "rectangle". If "buffered convex hull", then the the shape of the grid is a convex hull of the soundings, buffered by a distance of 'buffer'.}

\item{lon_0}{Longitude of projection center.}

\item{lat_0}{Latitude of projection center?}

\item{lat_1}{First standard parallel}

\item{lat_2}{Second standard parallel}
}
\value{
target_grid with cells of size res used for downscaling soundings
}
\description{
Uses the centroid of the union of the soundings as the projection center.
}
\examples{
grid <- generate_grid(soundings, res=330, buffer=2*330)
}
