% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmc_utils.R
\name{generate_qmc_unit_square}
\alias{generate_qmc_unit_square}
\title{Generate Sobol QMC Points in the Unit Square}
\usage{
generate_qmc_unit_square(n_points)
}
\arguments{
\item{n_points}{Number of QMC points to generate}
}
\value{
An n_points × 2 matrix of QMC points
}
\description{
Returns a matrix of quasi-Monte Carlo (QMC) points in the unit square \code{[0,1]^2}
using a Sobol sequence, suitable for mapping into triangle or quad regions.
}
\examples{
qmc <- generate_qmc_unit_square(128)
plot(qmc, asp = 1, main = "Sobol QMC Points in [0,1]^2")

}
