% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{grid_search_cv_stacked}
\alias{grid_search_cv_stacked}
\title{Parallel Grid Search CV for Stacked Spatial Models (joint grid)}
\usage{
grid_search_cv_stacked(
  prepared_list,
  hyperparameter_grid,
  folds = 5,
  metric = function(truth, pred) sqrt(mean((truth - pred)^2)),
  n_workers = future::availableCores(),
  use_progress = TRUE,
  verbose = TRUE,
  orthogonalize = NULL
)
}
\arguments{
\item{prepared_list}{A list of prepared_spatial_data objects.}

\item{hyperparameter_grid}{A data.frame of joint hyperparameter combinations. Each column should be named as \code{fieldname.param}, e.g., \code{water.rho}, \code{land.tau}.}

\item{folds}{Number of folds (integer) or list of test indices.}

\item{metric}{Function for error metric.}

\item{n_workers}{Number of parallel workers.}

\item{use_progress}{Show progress bar.}

\item{verbose}{Print messages.}

\item{orthogonalize}{Optional. A string indicating which field to orthogonalize.}
}
\value{
Data frame of grid results (mean error per combination).
}
\description{
Parallel Grid Search CV for Stacked Spatial Models (joint grid)
}
