% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate_region_list_batched_parallel}
\alias{integrate_region_list_batched_parallel}
\title{Batched Parallel Integrate Region List}
\usage{
integrate_region_list_batched_parallel(
  region_list,
  rule,
  basis,
  n_batches = 3000,
  n_workers = future::availableCores(),
  use_progress = TRUE
)
}
\arguments{
\item{region_list}{List of regions to integrate.}

\item{rule}{Integration rule object.}

\item{basis}{Basis object.}

\item{n_batches}{Number of batches (default: 3000).}

\item{n_workers}{Number of workers (default: all available cores).}

\item{use_progress}{Show progress bar (default: TRUE).}
}
\value{
Matrix of results (length(region_list) × basis$k).
}
\description{
Processes regions in batches to reduce per-task overhead
while retaining dynamic load balancing.
}
