% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_rules.R
\name{integration_rule_centroid_gaussian}
\alias{integration_rule_centroid_gaussian}
\title{Centroid Gaussian Integration Rule}
\usage{
integration_rule_centroid_gaussian(grid_centroids, params_list)
}
\arguments{
\item{grid_centroids}{A matrix of grid cell centroids (n_cells × 2).}

\item{params_list}{A list of parameter vectors per observation (mu_x, mu_y, sig_x, sig_y, angle).}
}
\value{
An integration rule object.
}
\description{
Creates an integration rule that computes Gaussian weights over grid centroids.
}
