% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{make_basis}
\alias{make_basis}
\title{Create a Spatial Basis Object}
\usage{
make_basis(evaluate_fn, k, crs = NULL, metadata = list())
}
\arguments{
\item{evaluate_fn}{A function of the form \code{f(coords)} that returns a matrix of basis
evaluations. \code{coords} should be a matrix or data frame with two columns (x and y).
The function must return a matrix of dimension \verb{[nrow(coords), k]}.}

\item{k}{An integer specifying the number of basis functions (i.e., the number of columns
in the matrix returned by \code{evaluate_fn}).}

\item{crs}{Optional. A coordinate reference system, either as an integer EPSG code,
a character string (e.g., \code{"EPSG:4326"}), or an \code{sf::st_crs()} object. This defines
the coordinate system in which the basis is constructed and evaluated. If \code{NULL},
the basis is assumed to be unitless (e.g., defined on [0,1]^2).}

\item{metadata}{An optional list of additional metadata, such as kernel parameters,
construction logs, or user-defined tags. Ignored by default methods.}
}
\value{
A list of class \code{"basis"} with fields:
\describe{
\item{evaluate_fn}{The user-defined function for evaluating basis functions at input coordinates.}
\item{k}{The number of basis functions.}
\item{crs}{The coordinate reference system (or \code{NULL} for unitless bases).}
\item{metadata}{A list of optional metadata elements.}
}
}
\description{
Constructs a basis object used to evaluate spatial basis functions for downscaling.
A basis is defined by a function that maps coordinates to a matrix of basis evaluations,
the number of basis functions used (\code{k}), an optional coordinate reference system (\code{crs}),
and additional optional metadata.
}
\examples{
# Example: basis defined on the unit square with two basis functions
my_eval <- function(coords) {
  cbind(sin(pi * coords[, 1]), cos(pi * coords[, 2]))
}
b <- make_basis(my_eval, k = 2, crs = NA)
b$evaluate_fn(matrix(c(0.5, 0.5), ncol = 2))

}
