% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{make_constant_basis}
\alias{make_constant_basis}
\title{Constant Basis Function}
\usage{
make_constant_basis(k = 1, crs = NULL, metadata = list())
}
\arguments{
\item{k}{Number of basis functions. Defaults to 1.}

\item{crs}{Optional coordinate reference system.}

\item{metadata}{Optional list of metadata.}
}
\value{
A \code{basis} object with class \code{constant_basis} and \code{basis}.
}
\description{
Returns a spatial basis that evaluates to 1 for all locations.
Used for testing performance of integration routines.
}
