% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_flat_prior}
\alias{make_flat_prior}
\title{Flat Prior (No Penalty)}
\usage{
make_flat_prior(basis)
}
\arguments{
\item{basis}{A basis object (used only to get \code{k})}
}
\description{
Flat Prior (No Penalty)
}
