% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_nystrom_prior}
\alias{make_nystrom_prior}
\title{Make a Nyström Prior}
\usage{
make_nystrom_prior(basis, tau = 1)
}
\arguments{
\item{basis}{A basis object created by \code{make_nystrom_basis()}}

\item{tau}{A global precision multiplier (default = 1)}
}
\description{
Uses eigenvalues from the Nyström basis as diagonal precisions
}
