% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{make_region_composite}
\alias{make_region_composite}
\title{Wrap a list of triangle geometries into a composite region}
\usage{
make_region_composite(triangles)
}
\arguments{
\item{triangles}{A list of triangle sf geometries}
}
\value{
A region_composite object
}
\description{
Wrap a list of triangle geometries into a composite region
}
