% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_ridge_prior}
\alias{make_ridge_prior}
\title{Ridge Prior (λ * I Penalty)}
\usage{
make_ridge_prior(basis, tau = 1)
}
\arguments{
\item{basis}{A basis object}

\item{tau}{Ridge penalty parameter}
}
\description{
Ridge Prior (λ * I Penalty)
}
