% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_sar_prior}
\alias{make_sar_prior}
\title{Make a SAR Prior}
\usage{
make_sar_prior(
  basis,
  rho = 0.99,
  tau = 1,
  W = NULL,
  scaling_weights = NULL,
  adjacency = "queen"
)
}
\arguments{
\item{basis}{A \code{geometry_basis} object.}

\item{rho}{Spatial dependence parameter (|rho| <= 1).}

\item{tau}{Precision scaling parameter (>0).}

\item{W}{Optional neighborhood matrix. If NULL, computed from basis.}

\item{scaling_weights}{Optional variance scaling vector (length matching the number of geometries).}
}
\value{
An object of class c("sar", "prior").
}
\description{
Constructs a spatial autoregressive (SAR) prior specification.
}
