% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_field.R
\name{make_spatial_field}
\alias{make_spatial_field}
\title{Specify a Spatial Random Field Term}
\usage{
make_spatial_field(
  basis = NULL,
  prior = NULL,
  integration_rule = NULL,
  label = "Spatial Field"
)
}
\arguments{
\item{basis}{Optional. A basis object created by \code{\link[=make_basis]{make_basis()}}, \code{\link[=make_geometry_basis]{make_geometry_basis()}}, etc.}

\item{prior}{Optional. A prior object created by e.g. \code{\link[=make_sar_prior]{make_sar_prior()}}.}

\item{integration_rule}{Optional. An integration rule object (e.g., from \code{\link[=integration_rule_qmc_average]{integration_rule_qmc_average()}}).}

\item{label}{Optional. A descriptive label for this spatial field (e.g., "Random effect").}
}
\value{
An object of class \code{"spatial_field"}.
}
\description{
Creates an object representing a latent spatial random field for downscaling models.
You can optionally supply a basis, prior, and integration rule.
If any are left NULL, defaults will be inferred inside \code{downscale()}.
}
\examples{
# Empty (defaults filled later)
sf1 <- make_spatial_field()

# With a geometry basis and SAR prior
grid <- sf::st_make_grid(sf::st_bbox(c(xmin=0, ymin=0, xmax=10, ymax=10)), n = c(5,5))
b <- make_geometry_basis(sf::st_sf(geometry=grid))
p <- make_sar_prior(b)
sf2 <- make_spatial_field(basis=b, prior=p, label="SAR spatial field")

}
