% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmc_utils.R
\name{map_unit_square_to_quad}
\alias{map_unit_square_to_quad}
\title{Map unit square points to a quadrilateral using bilinear interpolation}
\usage{
map_unit_square_to_quad(qmc_points, quad_coords)
}
\arguments{
\item{qmc_points}{An n × 2 matrix of points in \code{[0,1]^2}}

\item{quad_coords}{A 4 × 2 matrix of quad vertex coordinates (X, Y)}
}
\value{
An n × 2 matrix of points mapped into the quadrilateral
}
\description{
Maps an n × 2 matrix of points from the unit square \code{[0,1]^2} into a quadrilateral
defined by 4 coordinates, ordered counterclockwise: bottom-left, bottom-right,
top-right, top-left.
}
