% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{number2string}
\alias{number2string}
\title{Title}
\usage{
number2string(num, multiplier = 1, pad = FALSE, digits = 3)
}
\arguments{
\item{num}{number to be converter to string}

\item{multiplier}{premultiplier on num for string format}

\item{pad}{option to pad values with zeros for same length}

\item{digits}{number of digits in padded output. requires pad=TRUE>}
}
\value{
string value representing input number
}
\description{
Title
}
\examples{
padded_num <- number2string(0.1, multiplier=10, pad=TRUE)
}
