% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_helpers.R
\name{parse_downscale_formula}
\alias{parse_downscale_formula}
\title{Parse downscale formula}
\usage{
parse_downscale_formula(formula)
}
\arguments{
\item{formula}{A model formula}
}
\value{
A list:
\itemize{
\item fixed_terms: character vector of labels
\item random_bars: list of (1 | group) calls
\item spatial_terms: list of spatial_field objects
}
}
\description{
Parse downscale formula
}
