% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fitted_spatial_field}
\alias{plot.fitted_spatial_field}
\title{Plot a fitted spatial field with toggles}
\usage{
\method{plot}{fitted_spatial_field}(x, basemap = "OpenStreetMap", opacity = 0.7, ...)
}
\arguments{
\item{x}{A fitted_spatial_field object (from fit_spatial_field()).}

\item{basemap}{Character. One of "OpenStreetMap", "Esri.WorldImagery", or "CartoDB.Positron".}

\item{opacity}{Numeric between 0 and 1. Opacity of the grid polygons.}

\item{...}{Unused.}
}
\description{
Interactive leaflet visualization of the posterior mean and observations.
}
