% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.prior}
\alias{print.prior}
\title{Print Method for Prior Objects}
\usage{
\method{print}{prior}(x, ...)
}
\arguments{
\item{x}{A prior object created by e.g. make_sar_prior().}

\item{...}{Ignored.}
}
\description{
Nicely formats a prior object, showing its type, parameters, and dimension.
}
