% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region_composite}
\alias{region_composite}
\title{Create a composite region (e.g., triangulated polygon)}
\usage{
region_composite(parts, areas)
}
\arguments{
\item{parts}{A list of region_atomic() objects}

\item{areas}{A numeric vector of the same length as parts}
}
\value{
An S3 object of class "region_composite"
}
\description{
Create a composite region (e.g., triangulated polygon)
}
