% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_distributions.R
\name{scale_sf_geometries}
\alias{scale_sf_geometries}
\title{Scale every geometry in an sf object around its centroid}
\usage{
scale_sf_geometries(sf_obj, scale_factor)
}
\arguments{
\item{sf_obj}{An sf object with one or more rows}

\item{scale_factor}{Numeric scaling factor}
}
\value{
An sf object with all geometries scaled
}
\description{
Scale every geometry in an sf object around its centroid
}
\examples{
library(sf)
# Create example sf object with 2 rectangles
rects <- st_sf(
  id = 1:2,
  geometry = st_sfc(
    st_polygon(list(rbind(c(0,0), c(1,0), c(1,1), c(0,1), c(0,0)))),
    st_polygon(list(rbind(c(2,2), c(3,2), c(3,3), c(2,3), c(2,2))))
  ),
  crs = 4326
)
# Scale all geometries by 2
rects_scaled <- scale_sf_geometries(rects, 2)
# Plot
plot(st_geometry(rects), border = "blue")
plot(st_geometry(rects_scaled), border = "red", add = TRUE)
}
