% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_distributions.R
\name{scale_single_geometry}
\alias{scale_single_geometry}
\title{Scale a single sfc geometry around its centroid}
\usage{
scale_single_geometry(geom, scale_factor)
}
\arguments{
\item{geom}{An sfc geometry of length 1}

\item{scale_factor}{Numeric scaling factor}
}
\value{
A scaled sfc geometry
}
\description{
Scale a single sfc geometry around its centroid
}
