% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_solvers.R
\name{solve_system}
\alias{solve_system}
\title{estimate_target_grid}
\usage{
solve_system(precision_matrix, Atb, linear_solver = c("cg", "pcg", "cholesky"))
}
\arguments{
\item{linear_solver}{}
}
\description{
estimate_target_grid
}
