% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_intersection.R
\name{st_fast_intersection}
\alias{st_fast_intersection}
\title{st_fast_intersection (stack overflow)}
\usage{
st_fast_intersection(x, y, ...)
}
\arguments{
\item{x}{An sf object, first parameter in sf:st_intersects}

\item{y}{An sf object, second parameter in sf:st_intersects}

\item{...}{other arguments for sf::st_intersects}
}
\value{
Geometry if the the intersection between x and subset of y
}
\description{
st_fast_intersection (stack overflow)
}
