% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_intersection.R
\name{st_parallel_intersection}
\alias{st_parallel_intersection}
\title{st_parallel_intersection}
\usage{
st_parallel_intersection(
  soundings,
  fine_grid,
  mc.cores = parallel::detectCores(logical = FALSE),
  n_groups = mc.cores
)
}
\arguments{
\item{soundings}{An sf object, The observed soundings}

\item{fine_grid}{An sf object, a fine grid on which we want to estimate SIF}

\item{mc.cores}{number of cores to using in mclapply}

\item{n_groups}{number core groups. defaults to mc.cores parameter}
}
\value{
sf dataframe of intersections.
}
\description{
Similar to st_intersection, but runs in parallel to make use of more processors.
Uses the future pacakge (for compatibility with windows systems)
}
\examples{
if(interactive()) intersections <- st_parallel_intersection(soundings, fine_grid,mc.cores = 2)
}
