% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{triangulate_sf}
\alias{triangulate_sf}
\title{Triangulate an sf object into individual triangle geometries}
\usage{
triangulate_sf(geom)
}
\arguments{
\item{geom}{An sf geometry object (POLYGON or MULTIPOLYGON)}
}
\value{
An sfc collection of triangle geometries
}
\description{
Given an sf geometry or sf object, returns a list of triangle geometries.
Polygons that are already triangles are returned as-is.
}
