% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{wrap_atomic}
\alias{wrap_atomic}
\title{Wrap a triangle geometry into a region_atomic object}
\usage{
wrap_atomic(tri)
}
\arguments{
\item{tri}{A triangle sf geometry (must be exactly 3 unique exterior points)}
}
\value{
A region_atomic object
}
\description{
Wrap a triangle geometry into a region_atomic object
}
