% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{compute_precision.matern_fourier}
\alias{compute_precision.matern_fourier}
\title{Compute Precision Matrix for Matérn-Fourier Prior}
\usage{
\method{compute_precision}{matern_fourier}(prior)
}
\arguments{
\item{prior}{An object of class "matern_fourier".}
}
\value{
A sparse diagonal precision matrix.
}
\description{
Computes the diagonal precision matrix:
Q_j = tau * (kappa^2 + Laplacian_lambda_j)^alpha
The first entry (if constant_mode_included) may be overridden
}
