% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_matern_fourier_prior}
\alias{make_matern_fourier_prior}
\title{Make a Matérn-Fourier Prior}
\usage{
make_matern_fourier_prior(basis, tau = 1, kappa2 = 1, alpha = 1)
}
\arguments{
\item{basis}{A function_basis created by make_matern_fourier_basis()}

\item{tau}{Positive numeric, precision scaling parameter}

\item{kappa2}{Positive numeric, range parameter}

\item{alpha}{Positive numeric, smoothness parameter}
}
\value{
An object of class c("matern_fourier", "prior")
}
\description{
Constructs a prior over a Matérn-Fourier basis, parameterized by:
tau: Precision scaling parameter (>0)
kappa: Range parameter (>0)
alpha: Smoothness parameter (>0)
}
\details{
The constant mode precision (first entry) is recomputed as (kappa^2)^alpha if included.
}
