% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_field.R
\name{orthogonalize_design_block}
\alias{orthogonalize_design_block}
\title{Orthogonalize a design block against others}
\usage{
orthogonalize_design_block(stacked, target_label)
}
\arguments{
\item{stacked}{A list returned by \code{stack_spatial_design_matrices()}.}

\item{target_label}{A string indicating which field to orthogonalize (e.g., "Random effect").}
}
\value{
A modified version of \code{stacked}, with the target block replaced by an orthogonalized version.
}
\description{
Given a stacked spatial design object, modifies the design matrix so that
the block corresponding to \code{target_label} is orthogonal to all others.
}
